angel

+++ PYTHON/Ejercicios basicos/ejercicio3.2.py
@@ -1,5 +1,5 @@
 nombreFruta = input ("dime el nombre de una fruta: ")
-precioKilo = int(input ( " dime el precio de " +nombreFruta+ " por kilo: "))
-cantFruta = int (input ("dime cuantos kilos vas a querer: "))
+precioKilo = float(input ( " dime el precio de " +nombreFruta+ " por kilo: "))
+cantFruta = float (input ("dime cuantos kilos vas a querer: "))
 res=precioKilo*cantFruta
 print("eso es un total de" ,res, " euros.")
